consolidate |tcx|
consolidate function to |tcx| group
